requirements.txt
google-generativeai==0.3.2
python-telegram-bot==20.7
python-dotenv==1.0.0
requests==2.31.0
flask==3.0.0
flask-cors==4.0.0
PyPDF2==3.0.1
beautifulsoup4==4.12.2
pillow==10.0.1
lxml==4.9.3
apscheduler==3.10.4
yt-dlp==2024.8.6
numpy==1.24.3
python-dateutil==2.8.2
tweepy==4.14.0
facebook-sdk==3.1.0
schedule==1.2.0
gunicorn==21.2.0
openai==1.3.0
